+++ internal/glow/glow_darwin_embed.go
@@ -6,3 +6,5 @@ import _ "embed"
 
 //go:embed macos/bin/glow-macos
 var helperBinary []byte
+
+func init() { HelperEmbedded = true }